+++ oidc-idp/src/main/java/cz/muni/ics/oidc/server/claims/sources/EntitlementSource.java
@@ -4,9 +4,10 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.google.common.net.UrlEscapers;
-import cz.muni.ics.oidc.server.claims.ClaimSource;
+import cz.muni.ics.oidc.models.Facility;
 import cz.muni.ics.oidc.server.claims.ClaimSourceInitContext;
 import cz.muni.ics.oidc.server.claims.ClaimSourceProduceContext;
+import org.springframework.util.StringUtils;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -20,7 +21,8 @@ import java.util.Set;
  *         <b>custom.claim.[claimName].source.forwardedEntitlements</b> - forwardedEntitlements attribute name,
  *         if not specified, the forwarded entitlements will not be added to the list
  *     </li>
- *     <li><b>custom.claim.[claimName].source.capabilities</b> - capabilities attribute name</li>
+ *     <li><b>custom.claim.[claimName].source.resourceCapabilities</b> - resource capabilities attribute name for resources</li>
+ *     <li><b>custom.claim.[claimName].source.facilityCapabilities</b> - resource capabilities attribute name for facility</li>
  *     <li><b>custom.claim.[claimName].source.prefix</b> - string to be prepended to the value,</li>
  *     <li>
  *         <b>custom.claim.[claimName].source.authority</b> - string to be appended to the value, represents authority
@@ -33,15 +35,17 @@ import java.util.Set;
  */
 public class EntitlementSource extends GroupNamesSource {
 
-	private String eduPersonEntitlement;
-	private String capabilities;
-	private String prefix;
-	private String authority;
+	private final String eduPersonEntitlement;
+	private final String resourceCapabilities;
+	private final String facilityCapabilities;
+	private final String prefix;
+	private final String authority;
 
 	public EntitlementSource(ClaimSourceInitContext ctx) {
 		super(ctx);
 		eduPersonEntitlement = ctx.getProperty("forwardedEntitlements", null);
-		capabilities = ctx.getProperty("capabilities", null);
+		resourceCapabilities = ctx.getProperty("resourceCapabilities", null);
+		facilityCapabilities = ctx.getProperty("facilityCapabilities", null);
 		prefix = ctx.getProperty("prefix", null);
 		authority = ctx.getProperty("authority", null);
 	}
@@ -53,6 +57,11 @@ public class EntitlementSource extends GroupNamesSource {
 		JsonNodeFactory factory = JsonNodeFactory.instance;
 		ArrayNode result = new ArrayNode(factory);
 
+		Facility facility = null;
+		if (pctx.getClient() != null) {
+			facility = pctx.getPerunConnector().getFacilityByClientId(pctx.getClient().getClientId());
+		}
+
 		if (groupNamesJson != null) {
 			ArrayNode groupNamesArrayNode = (ArrayNode) groupNamesJson;
 			Set<String> groupNames = new HashSet<>();
@@ -63,13 +72,21 @@ public class EntitlementSource extends GroupNamesSource {
 				result.add(wrapGroupNameToAARC(value));
 			}
 
-			if (pctx.getClient() != null) {
+			if (facility != null && !StringUtils.isEmpty(this.resourceCapabilities)) {
 				Set<String> resultCapabilities = pctx.getPerunConnector()
-						.getResourceCapabilities(pctx.getClient().getClientId(), groupNames, capabilities);
+						.getResourceCapabilities(facility, groupNames, resourceCapabilities);
 
 				for (String capability : resultCapabilities) {
 					result.add(wrapCapabilityToAARC(capability));
 				}
+			}
+		}
+
+		if (facility != null && !StringUtils.isEmpty(this.facilityCapabilities)) {
+			Set<String> resultCapabilities = pctx.getPerunConnector()
+					.getFacilityCapabilities(facility, facilityCapabilities);
+			for (String capability : resultCapabilities) {
+				result.add(wrapCapabilityToAARC(capability));
 			}
 		}